fix: remove `common` & (some) `extension` circular imports
