+++ packages/extension/package.json
@@ -40,10 +40,11 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "yarn lint && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "e2e": "jest"
+    "e2e": "jest",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx"
   },
   "eslintConfig": {
     "extends": [
@@ -70,8 +71,14 @@
     "@types/react-custom-scrollbars": "^4.0.10",
     "@types/react-dom": "^17.0.0",
     "@types/react-sidebar": "^3.0.2",
-    "esbuild": "^0.14.34",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
+    "crypto-browserify": "^3.12.0",
+    "esbuild": "^0.14.38",
     "esbuild-jest": "^0.5.0",
+    "eslint": "^8.14.0",
+    "eslint-plugin-import": "^2.26.0",
+    "expect-puppeteer": "^6.1.0",
     "jest": "^27.5.1",
     "jest-puppeteer": "^6.1.0",
     "puppeteer-core": "^13.5.2",
